Add user signup and login routes; update package.json and .gitignore
