+++ .gitignore
@@ -4,3 +4,5 @@ users.json
 node_modules/
 package-lock.json
 .env
+./.env
+*/.env
