+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "projectxb",
-  "version": "0.1.0",
+  "version": "0.1.1",
   "description": "backend for projectXF",
   "main": "index.js",
   "type": "commonjs",
@@ -11,12 +11,11 @@
   "author": "Pedro Accarini Barros de Azevedo",
   "license": "ISC",
   "dependencies": {
+    "@octokit/core": "^6.1.2",
+    "@octokit/rest": "^18.12.0",
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
     "express": "^4.21.1",
     "fs": "^0.0.1-security"
-  },
-  "devDependencies": {
-    "@octokit/core": "^6.1.2"
   }
 }
