+++ routes/root.js
@@ -1,0 +1,9 @@
+const express = require('express');
+const router = express.Router();
+const fs = require('fs');
+
+router.get('/', (req, res) => {
+  res.send('Working...');
+});
+
+module.exports = router;
