+++ routes/signup.js
@@ -1,0 +1,33 @@
+const express = require('express');
+const router = express.Router();
+const fs = require('fs');
+
+router.post('/', (req, res) => {
+  const { username, email, password } = req.body;
+  const newUser = { username, email, password, userFrameworks: [] };
+
+  fs.readFile('users.json', 'utf8', (err, data) => {
+    let users = [];
+    if (!err && data) {
+      users = JSON.parse(data);
+    }
+
+    const userExists = users.some(user => user.email === email);
+
+    if (userExists) {
+      res.status(400).send('User already exists');
+    } else {
+      users.push(newUser);
+
+      fs.writeFile('users.json', JSON.stringify(users, null, 2), err => {
+        if (err) {
+          res.status(500).send('Error saving user');
+        } else {
+          res.status(201).send('User signed up successfully');
+        }
+      });
+    }
+  });
+});
+
+module.exports = router;
